converters.document:ConverterDocumentToImage - move from .converter_image
