+++ convertool/converters/converter_image.py
@@ -1,11 +1,8 @@
 from pathlib import Path
 from typing import ClassVar
 
-from acacore.models.file import File
-
 from convertool.util import TempDir
 
-from . import ConverterDocument
 from .base import ConverterABC
 
 
@@ -47,27 +44,6 @@ class ConverterImage(ConverterABC):
             tmp_dir.joinpath(dest_file.name).replace(dest_file)
 
         return [dest_file]
-
-
-class ConverterDocumentToImage(ConverterImage):
-    tool_names: ClassVar[list[str]] = ["document"]
-    platforms = list(set(ConverterImage.platforms or []).union(set(ConverterDocument.platforms or []))) or None
-    dependencies: ClassVar[list[str]] = [  # noqa: SIM222
-        *(ConverterImage.dependencies or []),
-        *(ConverterDocument.dependencies or []),
-    ] or None
-
-    def convert(self, output_dir: Path, output: str, *, keep_relative_path: bool = True) -> list[Path]:
-        output = self.output(output)
-        dest_dir: Path = self.output_dir(output_dir, keep_relative_path=keep_relative_path)
-
-        with TempDir(output_dir) as tmp_dir:
-            if not (pdfs := ConverterDocument(self.file, self.database, self.file.root).convert(tmp_dir, "pdf")):
-                return []
-
-            pdf = pdfs[0]
-
-            return ConverterPDFToImage(File.from_file(pdf, tmp_dir), self.database, tmp_dir).convert(dest_dir, output)
 
 
 class ConverterPDFToImage(ConverterImage):
